+++ src/lua/functions/creatures/player/mount_functions.cpp
@@ -27,7 +27,7 @@ int MountFunctions::luaCreateMount(lua_State* L) {
 	// Mount(id or name)
 	Mount* mount;
 	if (isNumber(L, 2)) {
-		mount = g_game().mounts.getMountByID(getNumber<uint32_t>(L, 2));
+		mount = g_game().mounts.getMountByID(getNumber<uint8_t>(L, 2));
 	} else if (isString(L, 2)) {
 		std::string mountName = getString(L, 2);
 		mount = g_game().mounts.getMountByName(mountName);
